requirements.txt
python-telegram-bot
google-generativeai
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
